requirements.txt
Django==3.0.3
djangorestframework==3.11.0
psycopg2==2.8.4
django-tenant-schemas
djangorestframework-simplejwt==4.4.0
django-filter==2.2.0
django-cors-headers==3.2.1
python-decouple==3.3